requirements.txt
gdown==4.7.1
lightning==2.1.3
nltk==3.8.1
pandas==2.1.4
Pillow==9.2.0
PyYAML==6.0.1
torch==2.1.2+cu118
torchvision==0.16.2
